Introduce Jinja2 templates for code generation
This change significantly refactors the way how code generation
works. Previously, pysmi code generator has been responsible for
producing a well-formed text document (e.g. JSON or pysnmp).

With this change, all code generations should be done through
Jinja2 templates by rendering them in the context of the parsed MIB
taking shape of the intermediate MIB representation which other parts
of pysmi provide.

By way of moving pysnmp code generation to Jinja2 template, the
Python code layout of pysnmp modules improved dramatically - it
is just one little step from being PEP8-compliant.

This template-based pysnmp code generator drops some backward
compatibility aids that keep Python MIB modules compatible with
older pysnmp versions. Perhaps in the followup patch we should
make the Python MIB module failing early and clearly on import
when it's old pysnmp importing it.
