+++ pysmi/codegen/templates/jsondoc.j2
@@ -1,0 +1,2 @@
+
+{{ mib|tojson(indent=2) }}
